pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "competitive-intel"
version = "1.0.0"
description = "LangGraph-based competitive intelligence pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-perplexity>=0.1.0",
    "perplexityai>=0.29.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "trafilatura>=1.12.0",
    "tavily-python>=0.5.0",
    "notion-client>=2.0.0",
]

[project.optional-dependencies]
postgres = [
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
include = ["competitive_intel*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
